11.25  setXY매개변수 변경 및 메인루프 순서정리 +u-> c -> e -> r

+++ Project1/Block.h
@@ -58,8 +58,8 @@ public:
 	void SetBlockXY(float, float);
 	void CopyBlock(short bl[4][4]);
 
-	short GetX();
-	short GetY();
+	float GetX();
+	float GetY();
 	short** PutBlock();
 	void RotateL();
 	void RotateR();
